feat: last push before deployment

+++ src/main/java/likelion14th/blog/service/ArticleService.java
@@ -4,6 +4,8 @@ import jakarta.persistence.EntityNotFoundException;
 import likelion14th.blog.domain.Article;
 import likelion14th.blog.dto.response.ArticleDetailResponse;
 import likelion14th.blog.dto.response.ArticleSummaryResponse;
+import likelion14th.blog.exception.ArticleNotFoundException;
+import likelion14th.blog.exception.*;
 import likelion14th.blog.repository.ArticleRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -25,14 +27,17 @@ public class ArticleService {
 
     @Transactional(readOnly = true)
     public ArticleDetailResponse getOneArticle(Long id) {
-        Article article = articleRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("해당 ID의 게시글을 찾을 수 없습니다."));
+        Article article = articleRepository.findById(id).orElseThrow(() -> new ArticleNotFoundException("해당 ID의 게시글을 찾을 수 없습니다."));
         return ArticleDetailResponse.from(article);
     }
 
     @Transactional //생략가능
-    public ArticleDetailResponse updateArticle(Long id, String title, String content) {
-        Article article = articleRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("해당 ID의 게시글을 찾을 수 없습니다."));
-
+    public ArticleDetailResponse updateArticle(Long id, String title, String content, String password) {
+        Article article = articleRepository.findById(id)
+                .orElseThrow(() -> new ArticleNotFoundException("해당 ID의 게시글을 찾을 수 없습니다."));
+        if(!article.getPassword().equals(password)){
+            throw new PermissionDeniedException("해당 게시글에 대한 수정 권한이 없습니다.");
+        }
         article.update(title, content);
         articleRepository.save(article);
         return ArticleDetailResponse.from(article);
@@ -50,9 +55,13 @@ public class ArticleService {
     }
 
     @Transactional
-    public Void deleteArticle(Long id){
-        articleRepository.deleteById(id);
+    public void deleteArticle(Long id, String password){
+        Article article = articleRepository.findById(id)
+                        .orElseThrow(() -> new ArticleNotFoundException("해당 ID의 게시글을 찾을 수 없습니다."));
 
-        return null;
+        if(!article.getPassword().equals(password)){
+            throw new RuntimeException("해당 게시글에 대한 삭제 권한이 없습니다.");
+        }
+        articleRepository.deleteById(id);
     }
 }
